+++ tests/Unit/V1/ErrorCollectionTest.php
@@ -33,7 +33,7 @@ class ErrorCollectionTest extends TestCase
     /**
      * @setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         $this->setUpManagerMock();