+++ tests/Unit/V1/ResourceNullTest.php
@@ -34,7 +34,7 @@ class ResourceNullTest extends TestCase
     /**
      * @setup
      */
-    public function setUp()
+    public function setUp(): void
     {
         $this->setUpManagerMock();