+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Changed
 
+- Support for PHP 5.6, 7.0 and 7.1 dropped, PHP ^7.2 is now required
 - **BREAKING**: Providing the fields `type` or `id` in a resource not as a string throws a `\Art4\JsonApiClient\Exception\ValidationException`, provide them always as strings instead
 
 ### Removed